♻️ Move open.go, conn.go, connector.go, driver.go and dsn_connector.go to internal/wrap and add aliases for them
