+++ internal/wrap/driver_test.go
@@ -1,4 +1,4 @@
-package sqlslog
+package wrap
 
 import (
 	"bytes"
