+++ internal/wrap/dsn_connector_test.go
@@ -1,4 +1,4 @@
-package sqlslog
+package wrap
 
 import (
 	"context"
