+++ internal/wrap/opts_aliases_test.go
@@ -12,8 +12,4 @@ const (
 	DurationString       = opts.DurationString
 )
 
-var (
-	DurationAttrFunc = opts.DurationAttrFunc
-	NewOptions       = opts.NewOptions
-	NewJSONHandler   = opts.NewJSONHandler
-)
+var NewJSONHandler = opts.NewJSONHandler
